+++ Content/Projectiles/Ranger/SuperFlare.cs
@@ -65,7 +65,7 @@ namespace RemnantOfTheAncientsMod.Content.Projectiles.Ranger
         }
         public override void OnHitNPC(NPC target, NPC.HitInfo hit, int damageDone)
         {
-            target.AddBuff(BuffID.OnFire, (int)Utils1.FormatTimeToTick(0, 0, 0, 10));
+            target.AddBuff(BuffID.OnFire, Utils1.FormatTimeToTick(0, 0, 0, 10));
         }
     }
 }
